Split io.py into several files on io submodule
Allow read_md3d_data to read specific datasets, reading all by default.
Allow to pass a different name of the parameters file.
