+++ modelling_earth/__init__.py
@@ -6,7 +6,8 @@ from .temperature import (
     empty_temperature_grid,
     litho_astheno_temperatures,
 )
-from .io import read_md3d_data, read_md3d_swarm
+from .io.grids import read_md3d_data
+from .io.swarm import read_md3d_swarm
 from .plot import plot_velocity_2d, plot_scalar_2d, save_plots_2d, plot_swarm_2d
