+++ modelling_earth/io/__init__.py
@@ -1,0 +1,3 @@
+"""
+Read and write data from/to files
+"""
