+++ modelling_earth/io/grids.py
@@ -1,0 +1,261 @@
+"""
+Read and write data structured in regular grids
+"""
+import os
+import numpy as np
+import xarray as xr
+import pandas as pd
+
+from .utils import _read_parameters, _read_times, PARAMETERS_FILE
+
+BASENAMES = {
+    "temperature": "Temper",
+    "density": "Rho",
+    "radiogenic_heat": "H",
+    "viscosity_factor": "Geoq",
+    "strain": "strain",
+    "viscosity": "visc_",
+    "velocity": "Veloc_fut",
+}
+DATASETS = (
+    "temperature",
+    "density",
+    "radiogenic_heat",
+    "viscosity_factor",
+    "strain",
+    "viscosity",
+    "velocity",
+)
+# Define which datasets are scalars meassured on the nodes of the grid, e.g. velocity is
+# not a scalar, and viscosity is meassured on the center of the finite elements
+SCALARS_ON_NODES = DATASETS[:5]
+
+
+def read_md3d_data(path, parameters_file=PARAMETERS_FILE, datasets=DATASETS):
+    """
+    Read the MD3D output files
+
+    Parameters
+    ----------
+    path : str
+        Path to the folder where the MD3D output files are located.
+    parameters_file : str (optional)
+        Name of the parameters file. It must be located inside the ``path`` directory.
+        Default to ``"param_1.5.3.txt"``.
+    datasets : tuple (optional)
+        Tuple containing the datasets that wants to be loaded.
+        The available datasets are:
+            - ``temperature``
+            - ``density"``
+            - ``radiogenic_heat``
+            - ``viscosity_factor``
+            - ``strain``
+            - ``viscosity``
+            - ``velocity``
+        By default, every dataset will be read.
+
+    Returns
+    -------
+    dataset :  :class:`xarray.Dataset`
+        Dataset containing data generated by MD3D.
+    """
+    # Read parameters
+    parameters = _read_parameters(os.path.join(path, parameters_file))
+    # Build coordinates
+    shape = tuple(parameters[i] for i in ("nx", "ny", "nz"))
+    coordinates = _build_coordinates(
+        max_coords=tuple(parameters[i] for i in ("x_max", "y_max", "z_max")),
+        shape=shape,
+    )
+    # Get array of times and steps
+    steps, times = _read_times(path, parameters["print_step"], parameters["stepMAX"])
+    # Create the coordinates dictionary containing the coordinates of the nodes
+    # and the time and step arrays. Then create data_vars dictionary containing the
+    # desired scalars datasets.
+    dims = ("time", "x", "y", "z")
+    coords = {
+        "time": times,
+        "step": ("time", steps),
+        "x": coordinates[0],
+        "y": coordinates[1],
+        "z": coordinates[2],
+    }
+    data_vars = {
+        scalar: (dims, _read_scalars(path, shape, steps, quantity=scalar))
+        for scalar in DATASETS
+        if scalar in SCALARS_ON_NODES
+    }
+
+    # Read velocity if needed
+    if "velocity" in datasets:
+        velocity_x, velocity_y, velocity_z = _read_velocity(path, shape, steps)
+        data_vars["velocity_x"] = (dims, velocity_x)
+        data_vars["velocity_y"] = (dims, velocity_y)
+        data_vars["velocity_z"] = (dims, velocity_z)
+
+    # Read viscosity if needed
+    if "viscosity" in datasets:
+        dims = ("time", "x_center", "y_center", "z_center")
+        # Add center coordinates of the finite elements to the coords
+        coords["x_center"] = (coordinates[0][1:] + coordinates[0][:-1]) / 2
+        coords["y_center"] = (coordinates[1][1:] + coordinates[1][:-1]) / 2
+        coords["z_center"] = (coordinates[2][1:] + coordinates[2][:-1]) / 2
+        # Add viscosity values located on the center of the finite elements
+        viscosity = _read_viscosity(path, steps, parameters)
+        data_vars["viscosity"] = (dims, viscosity)
+
+    return xr.Dataset(data_vars, coords=coords, attrs=parameters)
+
+
+def _build_coordinates(max_coords, shape):
+    """
+    Create grid coordinates
+
+    Parameters
+    ----------
+    max_coords : tuple
+        Maximum coordinates for each direction in the following order:
+        ``x_max``, ``y_max``, ``z_max`` in meters.
+    shape : tuple
+        Number of points for each direction in the following order:
+        ``nx``, ``ny``, ``nz``.
+
+    Returns
+    -------
+    coordinates : tuple
+        Tuple containing grid coordinates in the following order: ``x``, ``y``, ``z``.
+        All coordinates are in meters.
+    """
+    x_max, y_max, z_max = max_coords[:]
+    nx, ny, nz = shape[:]
+    # The z axis points downwards and the y axis is inverted to keep a right handed
+    # coordinate system.
+    x = np.linspace(0, x_max, nx)
+    y = np.linspace(-y_max, 0, ny)
+    z = np.linspace(-z_max, 0, nz)
+    return (x, y, z)
+
+
+def _read_scalars(path, shape, steps, quantity):
+    """
+    Read MD3D output scalar data
+
+    Read ``temperature``, ``density``, ``radiogenic_heat``, ``viscosity_factor`` and
+    ``strain``.
+
+    Parameters
+    ----------
+    path : str
+        Path to the folder where the MD3D output files are located.
+    shape: tuple
+        Shape of the expected grid.
+    steps : array
+        Array containing the saved steps.
+    quantity : str
+        Type of scalar data to be read.
+
+    Returns
+    -------
+    data: np.array
+        Array containing the MD3D output scalar data.
+    """
+    data = []
+    for step in steps:
+        filename = "{}_{}.txt".format(BASENAMES[quantity], step)
+        data_step = np.loadtxt(
+            os.path.join(path, filename), unpack=True, comments="P", skiprows=2
+        )
+        # Convert very small numbers to zero
+        data_step[np.abs(data_step) < 1.0e-200] = 0
+        # Reshape data_step
+        data_step = data_step.reshape(shape, order="F")
+        # Append data_step to data
+        data.append(data_step)
+    data = np.array(data)
+    return data
+
+
+def _read_velocity(path, shape, steps):
+    """
+    Read MD3D output velocity data
+
+    Parameters
+    ----------
+    path : str
+        Path to the folder where the MD3D output files are located.
+    shape: tuple
+        Shape of the expected grid.
+    steps : array
+        Array containing the saved steps.
+
+    Returns
+    -------
+    data: tuple of arrays
+        Tuple containing the components of the velocity vector.
+    """
+    velocity_x, velocity_y, velocity_z = [], [], []
+    for step in steps:
+        filename = "{}_{}.txt".format(BASENAMES["velocity"], step)
+        velocity = np.loadtxt(
+            os.path.join(path, filename), unpack=True, comments="P", skiprows=2
+        )
+        # Convert very small numbers to zero
+        velocity[np.abs(velocity) < 1.0e-200] = 0
+        # Separate velocity into their three components
+        velocity_x.append(velocity[0::3].reshape(shape, order="F"))
+        velocity_y.append(velocity[1::3].reshape(shape, order="F"))
+        velocity_z.append(velocity[2::3].reshape(shape, order="F"))
+    velocity_x = np.array(velocity_x)
+    velocity_y = np.array(velocity_y)
+    velocity_z = np.array(velocity_z)
+    return (velocity_x, velocity_y, velocity_z)
+
+
+def _read_viscosity(path, steps, parameters):
+    """
+    Read the viscosity files generated by MD3D. The viscosity values are defined on the
+    center of the finite elements instead of the nodes.
+
+    Parameters
+    ----------
+    path : str
+        Path to the folder where the MD3D output files are located.
+
+    Returns
+    -------
+    viscosity : data: np.array
+        Array containing the MD3D output scalar data.
+    """
+    # Determine the number of finite element centers per axes as the number of nodes per
+    # axes minus one
+    nx_centers, ny_centers, nz_centers = tuple(
+        parameters[i] - 1 for i in ("nx", "ny", "nz")
+    )
+    # List all viscosity_files
+    viscosity_files = [i for i in os.listdir(path) if BASENAMES["viscosity"] in i]
+    # Initialize the viscosity array
+    viscosity = np.zeros((steps.size, nx_centers, ny_centers, nz_centers), dtype=float)
+    # Fill the viscosity array with elements read from the data files
+    for step_index, step in enumerate(steps):
+        # Determine the rank value on the first step and check it for following steps
+        step_files = [
+            i
+            for i in viscosity_files
+            if "{}{}_".format(BASENAMES["viscosity"], step) in i
+        ]
+        if step_index == 0:
+            n_rank = len(step_files)
+        else:
+            if len(step_files) != n_rank:
+                raise ValueError(
+                    "Invalid number of ranks '{}' for step '{}'".format(
+                        len(step_files), step
+                    )
+                )
+        # Read rank file for this step and add the viscosity results to viscosity array
+        for rank in range(n_rank):
+            filename = "visc_{}_{}.txt".format(step, rank)
+            i, j, k, visc = np.loadtxt(os.path.join(path, filename), unpack=True)
+            i, j, k = tuple(indices.astype(int) for indices in (i, j, k))
+            viscosity[step_index, i, j, k] = visc
+    return viscosity
